requirements-dev.txt
Pillow
numpy
# evdev, spidev, lgpio excluded - Linux/Pi only
